+++ internal/storage/servers.go
@@ -20,3 +20,19 @@ func CreateServer(guildID string) error {
 
 	return nil
 }
+
+func DeleteServer(guildId string) error {
+	db, err := Database()
+
+	if err != nil {
+		return fmt.Errorf("failed to get database: %w", err)
+	}
+
+	_, err = db.Exec("DELETE FROM servers WHERE id = ?", guildId)
+
+	if err != nil {
+		return fmt.Errorf("failed to delete server: %w", err)
+	}
+
+	return nil
+}